// src/types/response/account.d.ts
import { DashBoardCommission, DashBoardPurchased, DashBoardRefs } from "./child";

type TLoginSuccessResponse = {
  id: number;
  email: string;
  address: any;
  token: string;
  name: any;
  refcode: string;
  status: number;
  ref: any;
  created_at: string;
  updated_at: string;
  password?: string;
};